refactor: dto 폴더 분류
