+++ src/main/java/com/catchtable/menu/dto/create/MenuCreateRequest.java
@@ -1,4 +1,4 @@
-package com.catchtable.menu.dto;
+package com.catchtable.menu.dto.create;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
